Separate getDedupedArticleReplies to make istanbul happy
