+++ lib/__tests__/getDedupedArticleReplies.test.js
@@ -1,4 +1,4 @@
-import { getDedupedArticleReplies } from './RelatedReplies';
+import getDedupedArticleReplies from '../getDedupedArticleReplies';
 
 describe('getDedupedArticleReplies', () => {
   it('handles when searchResults are not loaded yet', () => {
